refactor(ferment): split monolith into per-concern modules; resume UX overhaul
  Split src/extensions/ferment.ts (3077 lines) into 16 single-responsibility
  files under src/extensions/ferment/:

    index.ts            — extension factory + event handlers + commands (~890 lines)
    state.ts            — module state with typed accessors
    modes.ts            — isPlanMode/isExecMode/isAutoMode predicates
    colors.ts           — ANSI color helpers + truncateLabel/formatDuration
    format.ts           — formatFermentStatus/formatScopingContext/stripToolRefs
    judge.ts            — all 4 judge roles + computeFermentGrade
    worktree.ts         — worktree validation
    scoping.ts          — runScopingFlow + gate helpers
    nudge.ts            — maybeInjectAutoNudge + onStep/PhaseCompleted
    progress-overlay.ts — /progress 4-layer dialog
    tool-helpers.ts     — toolOk/toolErr + resolvePhase/resolveStep
    tool-schemas.ts     — TypeBox parameter schemas
    tools/lifecycle.ts  — create/list/scope/update/set_mode/complete
    tools/phases.ts     — activate/refine/complete/skip/fail
    tools/steps.ts      — start/complete/verify/skip/fail
    tools/knowledge.ts  — add_decision/add_memory

  Public API preserved — cli.ts and components/footer.ts updated to import
  from ./extensions/ferment/index.js (Node module resolution).

  Resume UX rewrite — explicit, never automatic:

  - Removed env-var-only auto-resume on session_start. The previous behavior
    (auto-mount widget, auto-flip paused→running, fire imperative nudge as
    soon as kimchi opened) was too aggressive. Now session_start only
    rehydrates if KIMCHI_ACTIVE_FERMENT is set in the current process tree
    (i.e. subagent inheritance) — never from session entries or storage scans.
  - Added resumeFerment() helper used by /ferment Continue picker and
    /ferment switch — flips paused→running, fires an imperative resume nudge
    ("RESUMING ferment X — pick up the work immediately, do NOT explain"),
    re-arms the scoping gate for drafts.
  - Engine messages updated: start_step now says "call start_step then spawn
    a subagent" (was "execute this step", which contradicted the planner's
    system prompt). Running-status step recovery message is explicit about
    the interrupted-mid-execution case.

  Widget cleanup — dashboard widget removed entirely:

  - Deleted buildDashboardWidget() and the persistent above-editor widget;
    /progress dropdown is now the only status surface.
  - Removed widget mount/unmount lifecycle from index.ts (~46 lines).
  - Removed widgetUpdateFn/setWidgetUpdateFn/refreshWidget from state.ts.
  - Removed the session turn counter (sessionTurnCount and friends) — was
    only consumed by the widget's context-budget warning.
  - Renamed widget.ts → colors.ts; kept ANSI helpers used by /progress.

  Bug fixes:

  - /progress no longer leaks a stuck widget into scrollback after Close.
    Previously /progress called setWidget(undefined) directly without
    updating widgetVisible, leaving the 30s ticker running and re-painting
    the widget at the wrong cursor position.
  - turn_end intercept now uses deliverAs: "followUp" on its sendUserMessage
    call. Without it, the harness throws "Agent is already processing.
    Specify streamingBehavior" when the user takes long enough to pick a
    dropdown option that the next turn is already starting.
